refactoring rate models
